[Java] Resolve all created publications in `CREATE_INGRESS_PUBLICATIONS` state to avoid errors upon closing invalid publications, i.e. those whose creation failed but get error was not polled.
This restores original logic which adds additional delay to the connection time, because we need to await all publications whereas a single connected publication would be enough (i.e. either we connect to the leader or receiver a redirect from a follower). However, to support response channels we would need to maintain all connections simultaneously anyhow.