Optimized register allocation aeron_atomic64_gcc_x86_64h (#1910)
There was a mixture of r and q registers being used.

q registers are less flexible because there are only 4:
rax...rdx
or their 32 bit counter parts
eax...edx

The `r` register are 16 and contain all the 'r' register
plus 12 extra.

So with more registers that can be used by the compiler,
there will be less register spilling.

Functionally there is no difference.