WIP fixes broken atomic instructions. (#1905)
I believe the curent atomic instructions are broken because no compiler
barrier is provided.

So although it can't happen that at the CPU level, loads and stores are
reorderd with the atomic instructions, the compiler still has the liberty
to move loads and stores.

Why do we need compiler barrier? Without the compiler barrier you would
not be able to do save handover to pass work from 1 thread to the next.

E.g. you set some values on a struct and then do e.g. an atomic increment of
a tail pointer, then it isn't guaranteed that the thread that sees the
incremented tail pointer, will see the updates to the struct.

The fix is to add memory clobber option `"memory", "cc"` to the instruction
to inform the compiler that that no load/stores before the atomic instruction
can jump after it, and no load/stores after the atomic instruction can jump
before it.

Co-authored-by: Dmytro Vyazelenko <696855+vyazelenko@users.noreply.github.com>